Fix TODO items 305-337: port renderer logic to match JS behavior
- TODO 305: Document u_time approach in M2RendererGL
- TODO 306: Fix bone_idx type to int16_t for negative sentinel detection
- TODO 307: Use dynamic geosetKey instead of hardcoded modelViewerGeosets
- TODO 308: Remove early return for zero-duration animations
- TODO 309: Fix M3RendererGL u_time from epoch to relative time
- TODO 310: Add model_matrix accessor to M3RendererGL
- TODO 311: Add non-const draw_calls accessor to M3RendererGL
- TODO 312: Fix getBoundingBox to not reject empty vertices
- TODO 313: Remove unused texture-ribbon.h include
- TODO 314: Implement reactive watcher block in MDXRendererGL
- TODO 315: Document setSlotFileLegacy as correct for string paths
- TODO 316: Fix MDXRendererGL u_time from epoch to relative time
- TODO 317: Verified setup_m2_separate_buffers default arg matches JS null
- TODO 318: Remove conditional guards on texture binding
- TODO 319: Document objectId optional handling
- TODO 320: Fix UV fallback reference
- TODO 321: Change WMODrawCall::count to uint32_t
- TODO 322: Verified rotation guard is correct
- TODO 323: Document polling approach
- TODO 324: Clear view arrays in dispose
- TODO 325: Add error logging for failed texture load
- TODO 326: Document isClassic check difference
- TODO 327: Add null check for updateAnimation
- TODO 328: Document view array value semantics
- TODO 329: Fix WMOLegacy texture wrap flags
- TODO 330: Document setSlotFileLegacy
- TODO 331: Document polling approach
- TODO 332: Change WMOLegacyDrawCall::count to uint32_t
- TODO 333: Fix updateGroups/updateSets to use local arrays
- TODO 334: Document dispose approach
- TODO 335: Add null check for updateAnimation
- TODO 336: Add error logging for null mpq
- TODO 337: Document out-of-bounds guard

Agent-Logs-Url: https://github.com/NathanNatNat/wow.export.cpp/sessions/ebc450c6-b243-4bd1-98b5-94ccd80c6ff4

+++ src/js/3D/renderers/WMOLegacyRendererGL.h
@@ -31,7 +31,7 @@ class M2LegacyRendererGL;
  */
 struct WMOLegacyDrawCall {
 	uint32_t start = 0;
-	uint16_t count = 0;
+	uint32_t count = 0;  // JS: unrestricted number — use uint32_t to avoid truncation
 	uint32_t blendMode = 0;
 	uint32_t material_id = 0;
 	int vertex_shader = 0;
